Break up slack adapter into multiple files
The single slack.rs file was starting to become a bit cluttered. Message
types and encoding logic was moved into slack/message.rs, and the
adapter implementation was moved into slack/mod.rs.
